+++ lib/markovian/tools.rb
@@ -1,6 +1,6 @@
 require 'markovian'
 require "markovian/tools/version"
-require "markovian/importers/twitter/csv_importer"
+require "markovian/importers/tweeter/csv_importer"
 require "markovian/publishers/tweeter"
 
 module Markovian
